Added now support for the new excitation input parameters.
Former-commit-id: 662dbf720a4e6fb26c448656cac5d89e5b0a9f4d

+++ src/data_structures/excitation.h
@@ -7,9 +7,9 @@ struct Excitation
 {
 	int type;
 	int label;
-	float emag;
-	float theta;
-	float phi;
+	double emag;
+	double theta;
+	double phi;
 	std::vector<int> ports;
 };
